move testable restapi server into place from rnd
